Store getFieldValue on ExecContext, rather than recreating it.
